# src/main/resources/static/json/swagger.json
{"swagger":"2.0","info":{"description":"更多Spring Boot相关文章请关注：http://xxxxblog.com/","version":"1.0","title":"Spring Boot中使用Swagger2构建RESTful APIs","termsOfService":"http://cccxxblog.com/","contact":{"name":"Henry"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"hello-controller","description":"Hello Controller"},{"name":"hello-world-controller","description":"Hello World Controller"},{"name":"user-controller","description":"User Controller"}],"paths":{"/hello/{name}/{age}":{"get":{"tags":["hello-world-controller"],"summary":"hello ....","description":"note: hello","operationId":"hellNameUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"name","in":"path","description":"名字","required":true,"type":"string","default":"nameX"},{"name":"age","in":"path","description":"年龄","required":true,"type":"integer","default":22,"format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"400":{"description":"test400"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/index/say":{"get":{"tags":["hello-controller"],"summary":"say","description":"note say","operationId":"sayUsingGET","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/users":{"get":{"tags":["user-controller"],"summary":"获取用户列表","operationId":"getUserListUsingGET","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/User"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["user-controller"],"summary":"创建用户","description":"根据User对象创建用户","operationId":"postUserUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"user","description":"用户详细实体user","required":true,"schema":{"$ref":"#/definitions/User"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/users/{id}":{"get":{"tags":["user-controller"],"summary":"获取用户详细信息","description":"根据url的id来获取用户详细信息","operationId":"getUserUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"id","description":"用户ID","required":true,"schema":{"$ref":"#/definitions/Long"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/User"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["user-controller"],"summary":"更新用户详细信息","description":"根据url的id来指定更新对象，并根据传过来的user信息来更新用户详细信息","operationId":"putUserUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"id","description":"用户ID","required":true,"schema":{"$ref":"#/definitions/Long"}},{"in":"body","name":"user","description":"用户详细实体user","required":true,"schema":{"$ref":"#/definitions/User"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["user-controller"],"summary":"删除用户","description":"根据url的id来指定删除对象","operationId":"deleteUserUsingDELETE","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"id","description":"用户ID","required":true,"schema":{"$ref":"#/definitions/Long"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}}},"definitions":{"User":{"type":"object","properties":{"age":{"type":"integer","format":"int32"},"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}}}}